+++ Crud.Api/Model/Brand/BrandDetailModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Crud.Api.Model.Brand
+{
+    public class BrandDetailModel : BrandModel
+    {
+        public int Id { get; set; }
+    }
+}
