+++ Crud.Api/Model/Brand/BrandModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Crud.Api.Model.Brand
+{
+    public class BrandModel
+    {
+        //public int? Id { get; set; }
+        public string Name { get; set; }
+        public string ShortDescription { get; set; }
+        public string Description { get; set; }
+    }
+}
