+++ Crud.Api/Model/Brand/UpdateBrandModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Crud.Api.Model.Brand
+{
+    public class UpdateBrandModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string ShortDescription { get; set; }
+        public string Description { get; set; }
+    }
+}
