--- Crud.Api/Model/BrandDetailModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Crud.Api.Model
-{
-	public class BrandDetailModel : BrandModel
-	{
-		public int Id { get; set; }
-	}
-}
