+++ Crud.Api/Model/Product/ProductDetailModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Crud.Api.Model.Product
+{
+    public class ProductDetailModel : ProductModel
+    {
+        public int Id { get; set; }
+
+    }
+}
