--- Crud.Api/Model/ProductDetailModel.cs
@@ -1,8 +1,0 @@
-﻿namespace Crud.Api.Model
-{
-	public class ProductDetailModel: ProductModel
-	{
-		public int Id { get; set; }
-	
-	}
-}
