--- Crud.Api/Model/ProductModel.cs
@@ -1,12 +1,0 @@
-﻿namespace Crud.Api.Model
-{
-	public class ProductModel
-	{
-		public string Title { get; set; }
-		public string StockCode { get; set; }
-		public decimal Price { get; set; }
-		public string Category { get; set; }
-		public string Gender { get; set; }
-		public bool IsActive { get; set; }
-	}
-}
