--- Crud.Api/Model/UpdateBrandModel.cs
@@ -1,10 +1,0 @@
-﻿namespace Crud.Api.Model
-{
-	public class UpdateBrandModel
-	{
-		public int Id { get; set; }
-		public string Name { get; set; }
-		public string ShortDescription { get; set; }
-		public string Description { get; set; }
-	}
-}
